:truck::recycle: Move permission classes into separate folder and change access modifiers where appropriate
