+++ Company.Domain/Permissions/NewsPermission.cs
@@ -1,0 +1,7 @@
+namespace Company.Domain
+{
+    public class NewsPermission : PermissionEntity
+    {
+        public new string Type { get; } = PermissionTypes.CanNews;
+    }
+}
