+++ Company.Domain/Permissions/RolesPermission.cs
@@ -1,0 +1,7 @@
+namespace Company.Domain
+{
+    public class RolesPermission : PermissionEntity
+    {
+        public new string Type { get; } = PermissionTypes.CanRoles;
+    }
+}
