+++ Company.Domain/Permissions/UsersPermission.cs
@@ -1,0 +1,7 @@
+namespace Company.Domain
+{
+    public class UsersPermission : PermissionEntity
+    {
+        public new string Type { get; } = PermissionTypes.CanUsers;
+    }
+}
